// src/categories/interfaces/category-info.enum.ts
export enum CategoryName {
    A = 'A',
    B = 'B',
    C = 'C',
    D = 'D',
    E = 'E'
}

export enum CategoryInfo {
    CATEGORY_A = 'Category A (World Elite - Top 10) → The top 10 players in the app ranking.',
    CATEGORY_B = 'Category B (High-Level Professionals) → 2,000+ points',
    CATEGORY_C = 'Category C (Advanced Competitors) → 1,001 - 2,000 points',
    CATEGORY_D = 'Category D (Growing Intermediate Players) → 501 - 1,000 points',
    CATEGORY_E = 'Category E (Beginners and New Players) → 200 - 500 points'
}